Add style.css and change images links to local
I added a style.css file in the root directory to make the home page
visually nice and cool.
Added images folder in the root directory and put some recipe images in it.
Also changed the image links to relative links instead of absolute links
so the loading time would be reduced.