Add style for the recipe pages
Created a style.css file for all the recipe gages in the recipes directory.
And linked the pages to the stylesheet.
Also added two buttons for Home and Next Recipe in the pages so viewer can switch pages more conveniently.
Used the same colors for the recipe pages as in the main page to keep the design consistent.